+++ CHANGELOG.md
@@ -29,6 +29,7 @@ v1.0.0 (??.??.????)
  * new class tx_rnbase_util_Typo3Classes to get TYPO3 classes independent of TYPO3 version
  * new Wrapper Class Tx_Rnbase_CommandLine (don't forget to load)  for \TYPO3\CMS\Core\Controller\CommandLineController or t3lib_cli which is compatible for all TYPO3 versions
  * Bugfix: tx_rnbase_util_Dates::date_mysql2tstamp liefert nun wieder NULL im Fehlerfall
+ * new Wrapper function tx_rnbase_util_Arrays::xml2array
 
 v0.15.18 (02.12.2015)
  * Bugfix: tx_rnbase_util_Dates::date_mysql2tstamp liefert nun immer einen Integer und nicht NULL im Fehlerfall